Fix bug with defaults on Ecto.Adapter.Queryable
Defaults are not saved on server-side when they don't create any
operations. This means we need to do a pass in the schema defaults
when querying the structure, and get the value from schema defaults
if its not set in the antidotec_map. The most notable occurence of
this bug happens with counter set as `0`.