Introduce top-level vx_server/vx_client apps and add WAL replication endpoint (#29)
Commit adds write-ahead-log replication functionality together with vx_server
and vx_client erlang applications. Vx_server serves as a top-level erlang
application for vaxine project. Vx_client is intended to be used to
communicating with vaxine from stand-alone applications.

vx_server and vx_client communicate through erlang terms over tcp for now. SSL
support and protobuf messaging to be added.

Write-ahead-log replication functionality is intended for replicating
materialized values from the antidote. This is the initial effort, and only
single partition is supported for now. Another limitation is that replication
always happens from the begging of the log, this to be changed in the future.
Replication that is streamed from the server does not have any back-pressure at
the moment, so vx_client is going to generate events as fast as it reads from
the socket.

On the Antidote site logging notification server concept was added to provide
notification stream to continuosly trigger replication. This could be replaced
by inotify-based notifications or other means in the future. disk_log:sync is
only triggered for commit log records, and not for all the records as it used to
be in the sync mode equal true.

VAX-59, VAX-60, VAX-61, VAX-68, VAX-73