[VAX-169, VAX-172] Add possibility to start WAL replication from specified position (#36)
Offset that is part of the WAL streaming api now could be passed to
vx_client:start_replication call to start replication from the specified
position. Such position should be treated as opaque and the onlye values
allowed that didn't originate from the streaming messages are 0 and `eof`

Commit introduces non-compatible change in Antidote, where update_start
operation is added to determine the first transaction's operation in
the WAL log.