VAX-490: Materialize transaction only when the cache has been updated up to global op id.
Fix in vx wal stream - previously global operation id check was always passing
if snapshot was stale. That has been fixed in vx_server application.

When receiving transactions from remote DCs global value of operation was not
updated, leading to non-monotonically increased values. Thus relying on that
value in vx_server was leading to errors. This was fixed, so that for remote dcs
operation number is combination of the monotonically increasing value within
local dc (global) and monotonically increasing value for transactions originated
in remote dc (local).